# server/api.yaml
openapi: 3.0.0
info:
  title: Code Improve API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 1.0
servers:
  - url: http://localhost:5000/
    description: Local server
components:
  securitySchemes:
    ApiTokenss:
      type: http
      scheme: bearer

    ApiKey:
      type: apiKey
      in: header
      name: apikey
paths:
...
  /api/auth/createuser:
    post:
      summary: Create a user
      tags:
       - User_Api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
                password:
                  type: string
              example:
                email: "test15686@gmail.com"
                name: "TesterFromSwagger"
                password: "test"
      responses:
        "201":
          description: Successfully Added to DB
  /api/auth/login:
    post:
      summary: Login a user
      tags:
       - User_Api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              example:
                email: "test15686@gmail.com"
                password: "test"
      responses:
        "201":
          description: User Logged in successfully
        "400":
          description: User not found
  /api/auth/getUserByEmail:
    post:
      summary: Find a user
      tags:
       - User_Api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              example:
                email: "postmantest1@gmail.com"
      responses:
        "201":
          description: Successfull
        "400":
          description: Not Found